renamed 'moduleContext' to 'context'
