+++ applications/standalone/src/model.ts
