+++ packages/client-core/src/index.ts
@@ -1,5 +1,4 @@
-export * from './common/model'
-export * from './common/modelContext'
+export * from './common/context'
 
 export * from './common/types'
 export * from './common/modules'
